chg: :recycle: make an independant node.py module
